requirements.txt
discord.py 
